requirements.txt
discord.py 
